+++ client/src/components/TestPage.vue
@@ -1,26 +1,25 @@
 <template>
-    <div>
-      <NavbarVue />
-      <CustomTestForm class="test" />
-    </div>
-  </template>
-  
-  <script>
-  import NavbarVue from './PageNavbar.vue';
-  import CustomTestForm from './CustomTestForm.vue';
-  
-  export default {
-    components: {
-      NavbarVue,
-      CustomTestForm,
-    },
-  };
-  </script>
-  
-  <style scoped>
-    .test {
-        margin-top: 30px;
-        margin-left: 30px;
-    }
-  </style>
-  
+  <div>
+    <NavbarVue />
+    <CustomTest class="test" />
+  </div>
+</template>
+
+<script>
+import NavbarVue from "./PageNavbar.vue";
+import CustomTest from "./CustomTest.vue";
+
+export default {
+  components: {
+    NavbarVue,
+    CustomTest,
+  },
+};
+</script>
+
+<style scoped>
+.test {
+  margin-top: 30px;
+  margin-left: 30px;
+}
+</style>
